+++ machine_learning/src/training/trainer.rs
@@ -8,7 +8,7 @@ use crate::{
 ///
 /// Either the training must go on or the last call was the last.
 pub struct TrainResult<'trainer> {
-    pub losses: &'trainer [f32],
+    pub losses: &'trainer [f64],
     pub was_last: bool,
 }
